SCRUM-97 updates with consumables
class CharacterPawn now takes shield value = 100(for now)

take_dmg() now checks if shield > 0 if so dmg is done to the shield. if not dmg goes to health

right now the shield takes dmg in 4 chunks of 25. this will be changed because enemies should be able to dmg shield for various amounts